+++ backend/migrations/20240815140308_party_lists.js
@@ -4,7 +4,7 @@
  */
 exports.up = function (knex) {
   return knex.schema.createTable("PARTY_LISTS", function (table) {
-    table.bigInteger("LIST_ID").primary(); // BIGINT primary key for list ID
+    table.increments("LIST_ID").primary(); // BIGINT primary key for list ID
     table.string("LIST_NAME", 100); // VARCHAR(100) for list name
     table.bigInteger("COUNT_OF_VOTES"); // BIGINT for count of votes
     table.boolean("IS_APROVED").defaultTo(false); // BOOLEAN with default value FALSE for approved status
